doorbell/requirements.txt
json_sempai
PyYAML
paho-mqtt
rpi-backlight
Kivy
twisted
pydantic_yaml[ruamel]
python-box[ruamel.yaml]
mypy
kivymd